Extract logical instructions to instructions_logical.go module
